% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewe_nwatlantic.R
\docType{data}
\name{ewe_nwatlantic_base}
\alias{ewe_nwatlantic_base}
\alias{ewe_nwatlantic_env}
\title{EWE data for the NW Atlantic}
\format{
\subsection{\verb{EWE NW Atlantic}}{

A tibble with 35640 rows and the following
8 columns:
\describe{
\item{file_name}{The path to the file name, note that the path could be a
relative path. This column is helpful for understanding which EWE model
and which run the data came from.
}
\item{type}{The type refers to what data type this row is, e.g., landings,
catch, biomass, weight, and mortality.
}
\item{year}{A four-digit integer specifying the year.}
\item{month}{A one- or two-digit integer specifying the month.}
\item{function_group}{A string specifying the functional group that this
row pertains to.
}
\item{value}{A real number containing the value of interest.}
\item{fleet}{An integer specifying which fleet the data pertains to.}
}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35640 rows and 8 columns.
}
\source{
\url{https://www.github.com/nmfs-ost/}
}
\usage{
ewe_nwatlantic_base

ewe_nwatlantic_env
}
\description{
Ecopath with Ecosim data for the Northwest Atlantic ocean, where each data
object is from a different run. E.g., \code{base} is from the base model run and
\code{env} is from the run using environmental variables.
}
\references{
Chagaris, D., Drew, K., Schueller, A., Cieri, M., Brito, J., and
Buchheister, A. 2020. Ecological reference points for Atlantic menhaden
established using an ecosystem model of intermediate complexity. Frontiers
in Marine Science, 7:606417. 10.3389/fmars.2020.606417.
}
\keyword{datasets}
