% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functional_groups.R
\name{get_functional_groups}
\alias{get_functional_groups}
\title{Get the functional groups from an EwE file}
\usage{
get_functional_groups(file_path)
}
\arguments{
\item{file_path}{The path to the EwE file.}
}
\value{
A vector of functional group names.
}
\description{
Functional group names are useful for the column names of output data from
an EeW model, and thus, this function is a way to get them automatically
from the EwE output. The functional group names come from the basic
estimates file.
}
\examples{
get_functional_groups(
  file_path = fs::path(
    system.file("extdata", package = "ecosystemdata"),
    "ewe_nwatlantic", "base_run", "basic_estimates.csv"
  )
)
}
