% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_.R
\name{load_csv_ewe}
\alias{load_csv_ewe}
\title{Load in EwE monthly output data}
\usage{
load_csv_ewe(file_path, model_years, functional_groups)
}
\arguments{
\item{file_path}{The path to the CSV file containing the EwE output data.}

\item{model_years}{A vector of years corresponding to the model years. TODO: this can be removed
after havgina a utility function to extract model years globally from annual data}

\item{functional_groups}{A vector of names of the functional groups in the model.}
}
\description{
The function loads in Ecopath with Ecosim (EwE) monthly output data from a CSV file.
}
\examples{
\dontrun{
# The following example is not run by default because these files are only
# included in the GitHub clone of the repository and not in the package data
load_csv_ewe(
  file_path = fs::path(
    "data-raw", "ewe_nwatlantic", "base_run", "biomass_monthly.csv"
  ),
  model_years = 1985:2017,
  functional_groups = c(
    "StripedBass0",
    "StripedBass2_5",
    "StripedBass6",
    "AtlanticMenhaden0",
    "AtlanticMenhaden1",
    "AtlanticMenhaden2",
    "AtlanticMenhaden3",
    "AtlanticMenhaden4",
    "AtlanticMenhaden5",
    "AtlanticMenhaden6",
    "SpinyDogfish",
    "BluefishJuvenile",
    "BluefishAdult",
    "WeakfishJuvenile",
    "WeakfishAdult",
    "AtlanticHerring0_1",
    "AtlanticHerring2",
    "Anchovies",
    "Benthos",
    "Zooplankton",
    "Phytoplankton",
    "Detritus"
  )
)
}
}
