% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_.R
\name{load_model}
\alias{load_model}
\title{Load an ecosystem model}
\usage{
load_model(..., type = c("ewe", "atlantis"))
}
\arguments{
\item{...}{Arguments that are passed onto lower level \verb{load_model_()*}
functions. Such as those needed for \code{load_model_ewe()}, which are
\code{directory} and \code{functional_groups}.}

\item{type}{A string indicating which type of model data you want to load.
The default is ewe. Strings
should be all lower case text.}
}
\value{
A tibble is returned that matches the structure of \link{ewe_nwatlantic_base}.
}
\description{
Load the necessary files from an ecosystem model and return a single, long
data frame of information.
}
