% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_environment_data.R
\name{reshape_environment_data}
\alias{reshape_environment_data}
\title{Reshape environmental data and return it as a tibble.}
\usage{
reshape_environment_data(index_name, values, years, lag, impacted_group)
}
\arguments{
\item{index_name}{A string indicating the name of the environmental index.}

\item{values}{A numeric vector of values corresponding to the environmental index.}

\item{years}{An integer vector of years corresponding to the values.}

\item{lag}{An integer indicating the lag applied to the environmental index.}

\item{impacted_group}{A string indicating the group impacted by the environmental index.}
}
\value{
A tibble containing the reshaped environmental data.
}
\description{
Reshape environmental data and return it as a tibble.
}
\examples{
data <- reshape_environment_data(
  index_name = "index1",
  values = rnorm(24, mean = 1, sd = 0.2),
  years = 1984:1985,
  lag = 1,
  impacted_group = "menhaden 0"
) |>
  dplyr::bind_rows(
    reshape_environment_data(
      index_name = "index2",
      values = rnorm(24, mean = 1, sd = 0.2),
      years = 1985:1986,
      lag = 0,
      impacted_group = "menhaden 1"
    )
  )

}
